implemented AuthGuard and used it for todos (still no JWT)

+++ backend/src/todos/todos.service.ts
@@ -3,7 +3,9 @@ import { CreateTodoDto } from './dto/create-todo.dto';
 import { UpdateTodoDto } from './dto/update-todo.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Todo } from './entities/todo.entity';
-import { Repository } from 'typeorm';
+import { IsNull, Repository } from 'typeorm';
+import { User } from 'src/users/entities/user.entity';
+import { CreateUserDto } from 'src/users/dto/create-user.dto';
 
 @Injectable()
 export class TodosService {
@@ -11,22 +13,31 @@ export class TodosService {
 
   create(createTodoDto: CreateTodoDto) {
     const newTodo = this.todoRepository.create(createTodoDto);
-    return this.todoRepository.save(newTodo);
+    return this.todoRepository.save(newTodo, {});
   }
 
-  findAll(): Promise<Todo[]> {
-    return this.todoRepository.find();
+  findAll(userID: number): Promise<Todo[]> {
+    return this.todoRepository.find({
+      select: ['id', 'name', 'checked'],
+      where: { user: { id: userID } },
+    });
   }
 
-  findOne(id: number): Promise<Todo> {
-    return this.todoRepository.findOneBy({ id });
+  findOne(id: number, userID: number): Promise<Todo> {
+    return this.todoRepository.findOneBy({ id, user: { id: userID } });
   }
 
-  update(id: number, updateTodoDto: UpdateTodoDto) {
-    return this.todoRepository.update({ id }, updateTodoDto);
+  update(id: number, userID: number, updateTodoDto: UpdateTodoDto) {
+    return this.todoRepository.update(
+      { id, user: { id: userID } },
+      updateTodoDto,
+    );
   }
 
-  remove(id: number) {
-    return this.todoRepository.delete({ id });
+  async remove(id: number, userID: number) {
+    return (await this.todoRepository.delete({ id, user: { id: userID } }))
+      .affected == 0
+      ? false
+      : true;
   }
 }
